ts implement, restracture
