+++ styles/V2/Contact/Contact.module.css
@@ -10,12 +10,3 @@
     font-weight: 600;
     color: #b7bccb;
 }
-
-.check {
-    margin-top: 30px !important;
-}
-
-/* font-family: 'Amiri',
-serif;
-font-family: 'Glory',
-sans-serif; */
